separated admin/volunteer jurisdition editors
