+++ client/web-admin/src/components/Dashboard/JurisdictionEditor/EditTable/Row.js
