+++ client/web-admin/src/components/Dashboard/JurisdictionEditor/EditTable/index.js
